LayoutTests/imported/w3c/web-platform-tests/css/css-nesting/top-level-parent-pseudo-specificity-expected.txt


FAIL CSS Nesting: Specificity of top-level '&' assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
